--- precode.go
@@ -1,90 +1,0 @@
-package main
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"sync"
-)
-
-// Generator генерирует последовательность чисел 1,2,3 и т.д. и
-// отправляет их в канал ch. При этом после записи в канал для каждого числа
-// вызывается функция fn. Она служит для подсчёта количества и суммы
-// сгенерированных чисел.
-func Generator(ctx context.Context, ch chan<- int64, fn func(int64)) {
-	// 1. Функция Generator
-	// ...
-}
-
-// Worker читает число из канала in и пишет его в канал out.
-func Worker(in <-chan int64, out chan<- int64) {
-	// 2. Функция Worker
-	// ...
-}
-
-func main() {
-	chIn := make(chan int64)
-
-	// 3. Создание контекста
-	// ...
-
-	// для проверки будем считать количество и сумму отправленных чисел
-	var inputSum int64   // сумма сгенерированных чисел
-	var inputCount int64 // количество сгенерированных чисел
-
-	// генерируем числа, считая параллельно их количество и сумму
-	go Generator(ctx, chIn, func(i int64) {
-		inputSum += i
-		inputCount++
-	})
-
-	const NumOut = 5 // количество обрабатывающих горутин и каналов
-	// outs — слайс каналов, куда будут записываться числа из chIn
-	outs := make([]chan int64, NumOut)
-	for i := 0; i < NumOut; i++ {
-		// создаём каналы и для каждого из них вызываем горутину Worker
-		outs[i] = make(chan int64)
-		go Worker(chIn, outs[i])
-	}
-
-	// amounts — слайс, в который собирается статистика по горутинам
-	amounts := make([]int64, NumOut)
-	// chOut — канал, в который будут отправляться числа из горутин `outs[i]`
-	chOut := make(chan int64, NumOut)
-
-	var wg sync.WaitGroup
-
-	// 4. Собираем числа из каналов outs
-	// ...
-
-	go func() {
-		// ждём завершения работы всех горутин для outs
-		wg.Wait()
-		// закрываем результирующий канал
-		close(chOut)
-	}()
-
-	var count int64 // количество чисел результирующего канала
-	var sum int64   // сумма чисел результирующего канала
-
-	// 5. Читаем числа из результирующего канала
-	// ...
-
-	fmt.Println("Количество чисел", inputCount, count)
-	fmt.Println("Сумма чисел", inputSum, sum)
-	fmt.Println("Разбивка по каналам", amounts)
-
-	// проверка результатов
-	if inputSum != sum {
-		log.Fatalf("Ошибка: суммы чисел не равны: %d != %d\n", inputSum, sum)
-	}
-	if inputCount != count {
-		log.Fatalf("Ошибка: количество чисел не равно: %d != %d\n", inputCount, count)
-	}
-	for _, v := range amounts {
-		inputCount -= v
-	}
-	if inputCount != 0 {
-		log.Fatalf("Ошибка: разделение чисел по каналам неверное\n")
-	}
-}
